% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{availableSpecies}
\alias{availableSpecies}
\title{List or check the available list of species.}
\usage{
availableSpecies(species = NULL, region = NULL, callFunction = NULL)
}
\arguments{
\item{species}{A character indicating one of the species name. The function is case insensitive.}

\item{region}{A character indicating one of the region names (depending on the species). The function is case insensitive. Default is NULL.}

\item{callFunction}{A character indicating the function caller. For internal use. Default is NULL.}
}
\value{
When called without parameters, return a matrix with all the available species names and region. If called internally by a function,
(in order to check the species and region), it will trough an error if the species and region are not available, otherwise the function will return nothing.
}
\description{
..
}
\details{
To be completed.
}
\examples{
availableSpecies()  # show all the species available.
availableSpecies('eucalyptus')   # check if it's available.
\dontrun{
# check if it's available. It doesn't exist, so an error will be triggered.
availableSpecies('test')
}
}
\seealso{
Other Internals: \code{\link{Dispatcher}},
  \code{\link{EstimateDBHSd}},
  \code{\link{GetBAAfterThinning}},
  \code{\link{ProjectBasalArea}},
  \code{\link{ProjectDominantHeight}},
  \code{\link{ProjectNTrees}}, \code{\link{ProjectVolume}},
  \code{\link{SiteIndex}}
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
\concept{Internals}
