% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{BatchGrowing}
\alias{BatchGrowing}
\title{Main Growing batch parser}
\usage{
BatchGrowing(data, ...)
}
\arguments{
\item{data}{A matrix with an id, Age, Hd, N, Dg, DBH_sd, BA, Vt, years, thinninAge, n.afterThin, thinTypes and thin_coeff
as columns (following the example)}

\item{...}{Arguments passed to \code{\link[fpfi3]{Growing}}.}
}
\value{
A data.frame with the Growing results
}
\description{
Parser to the \code{\link[fpfi3]{Growing}} function, for multiple simulations
}
\details{
To be completed.
}
\examples{
values <- c(1, 8.46, 30.813846, 350.0427, 28.08013, 22.043504, 250, 25, '9',
'245', "low", "1", 2, 7.571743, 27.388918, 378, 25.29584, 18.730381, 220, 25,
'8,10', '320.3,250', "low,PRD", "0.9,0.9", 3, 6.571986, 25.437695, 478.8546,
22.92591, 20.037553, 220, 25, '7,9', '387.1,250', 'low,PRD', "0.9,0.9", 4, 5.52,
22.5, 558.1, 20.6, 18.977, 180, 25,'6,7', '503.8,350', 'low,PRD', "0.9,0.9",
5, 1.621429, 9.792857, 456.7143, 10.20179, 25, 3.852857, 25, '5,7', '400,350',
'low,PRD', "0.9,0.9")
}
\seealso{
\code{\link[fpfi3]{Growing}}.
}
\author{
Ãlvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
