% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bbarplot}
\alias{bbarplot}
\title{Bar Plots Modification}
\usage{
bbarplot(df, legend_space = NULL, legend_offset = 0.4,
  text_rotation = 0, macrostands = NULL, operation = "harvest",
  ages = NULL, type = "value", ...)
}
\arguments{
\item{df}{A \code{\link[fpfi3]{FullSimulation}} object or a data.frame, data.table or matrix (The first column must be an identifier field
and the other columns are the variables to be pottled (for example log products).}

\item{legend_space}{The extra right space to allocate the legend. If NULL no extra space will be considerer. Default is NULL.}

\item{legend_offset}{The x axis position of the legend; valid ranges goes from -1 to 1 (see \code{inset}
in \code{\link[graphics]{legend}}.If NULL, no legend will be displayed. Default is 0.4 (at the right).}

\item{text_rotation}{The text rotation inside every bar (see srt in \code{\link{text}}). If NULL, no text will be displayed.
Default is 0 (no rotation).}

\item{macrostands}{A character vector indicating the macrostands to plot.
Only used if \code{df} is a \code{\link[fpfi3]{FullSimulation}} object.
Default is to use the first macrostand available.}

\item{operation}{'harvest' or 'thinning'. Default is 'harvest'.
Only used if \code{df} is a \code{\link[fpfi3]{FullSimulation}} object.}

\item{ages}{An integer speDefault is max age.
Only used if \code{df} is a \code{\link[fpfi3]{FullSimulation}} object.}

\item{type}{A character indicating what type of plot to display.
Only used if \code{df} is a \code{\link[fpfi3]{FullSimulation}} object. Options are:
\itemize{
\item \code{'logs'}: Number of logs per dbh class and type.
\item \code{'value'}: value per dbh class, adding all the value of the log type by dbh class.
\item \code{'rvalue'}: relative value per dbh class, the same at \code{'value'} but in % relative to the total dbh class.
\item \code{'volume'}: volume per dbh class, adding all the volume of the log type by dbh class.
In all these plots, the frequency of the class it's shown in small numbers below the title.
Defaul is 'value'.
}}

\item{...}{Aditional parameters passed to \code{\link[graphics]{barplot}} (as xlab, ylab, col, etc).}
}
\description{
Creates a modified bar plot, using the native \code{\link[graphics]{barplot}} as base.
}
\details{
Note that if \code{df} is a \code{\link[fpfi3]{FullSimulation}} object, you could generate some predefined
plots. Any arguments passed as \code{...} will have priority over any other internal pre-defined value.
}
\examples{
library(fpfi3)
library(data.table)

# We're going to use the package demo data:
# initial_conditions, taper_coeffs, demo_products
trees <- DoBucking(initial_conditions, taper_coeffs, demo_products,
                   id='id', dbh='Dg', ht='Hd',
                   hp='hp', hm='hm', model='dss', volform='smalian', buck='lmak',
                   output='logs')

# As we're using the initial conditions, we're gonna get few logs.
treesm <- trees$logs
treesm[, volval:=volume*value]

tlength <- dcast(treesm, id~product, length, value.var="volval")
tvol <- dcast(treesm, id~product, sum, value.var="volume")
tval <- dcast(treesm, id~product, sum, value.var="volval")

# Number of logs
bbarplot(tlength, legend_space=10, legend_offset=.35,
         xlab="Tree", ylab="Logs", ylim=c(0, 7),
         main="Number of logs by tree")

# Volume
bbarplot(tvol, legend_space=10, legend_offset=.35,
         xlab="Tree", ylab="Volume", ylim=c(0, 1),
         main="Volume per tree")

# Value
bbarplot(tval, legend_space=10, legend_offset=.35,
         xlab="Tree", ylab="Value", ylim=c(0, 16),
         main="Value by tree")

# Relative log value
bbarplot(cbind(tval[, 1], tval[, -1] / rowSums(tval[, -1])*100),
         legend_space=10, legend_offset=.35,
         xlab="Tree", ylab="\% of the tree value", ylim=c(0, 100),
         main="Value by tree")

}
\seealso{
\code{\link[fpfi3]{DoBucking}}, \code{\link[fpfi3]{FullSimulation}}
}
\author{
Ãlvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
