% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Dispatcher}
\alias{Dispatcher}
\title{Dispatcher function to select one of the growth internal functions according to the species and region.}
\usage{
Dispatcher(species, region, lista, callFunction = "", ...)
}
\arguments{
\item{species}{A character indicating the species.}

\item{region}{A character indicating the region/type.}

\item{lista}{A list of functions, with structure \code{list(species=list(region))}. If region is NULL, first item inside the species will be used.}

\item{callFunction}{A character indicating the function caller. For internal use. Default is NULL.}

\item{...}{Other arguments passed to the selected function.}
}
\description{
This function have been designed to be used as internal. The Site Index, DBH sd estimation,
height simulation... are the functions that utilize this dispatcher to select the appropiate parameters and sub-functions according to the
species and region
}
\details{
To be completed.
}
\seealso{
Other Internals: \code{\link{EstimateDBHSd}},
  \code{\link{GetBAAfterThinning}},
  \code{\link{ProjectBasalArea}},
  \code{\link{ProjectDominantHeight}},
  \code{\link{ProjectNTrees}}, \code{\link{ProjectVolume}},
  \code{\link{SiteIndex}}, \code{\link{availableSpecies}}
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
\concept{Internals}
\keyword{internal}
