% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bucking.R
\name{DoBucking}
\alias{DoBucking}
\title{Stem bucking simulation and merchantable volume calculation}
\usage{
DoBucking(data, coeffs, products, id = NULL, dbh = "dbh", ht = "ht",
  hp = NULL, hm = NULL, stump = 0.1, lseg0 = 0.2,
  hdr_range = c(0.5, 2.5), dbh_range = c(10, 60), ht_range = c(1.3,
  60), model = "dss", volform = "smalian", buck = "lmak",
  grade_wd = TRUE, mg_disc = 0, length_disc = 0,
  output = "summary", byClass = FALSE)
}
\arguments{
\item{data}{A data.frame, data.table or matrix, with colnames id, ht, and dbh at least. Every row corresponds to a tree.}

\item{coeffs}{A vector, with the parameters for the taper model. See \code{\link[fpfi3]{taper_selec}} for more information.}

\item{products}{A data.frame of 3 o 4 columns at least, depending on the \code{buck} option selected. See \code{\link[fpfi3]{lmak}}
and \code{\link[fpfi3]{lpBuck}} for more details.}

\item{id}{The column name in \code{data}, that corresponds to the input id. It will be returned with the ouput table. Default is NULL (will not be used).}

\item{dbh}{The name of the dbh column in \code{data}. Default is 'dbh'.}

\item{ht}{The name of the ht column in \code{data}. Default is 'ht'.}

\item{hp}{Prunning height. Could be NULL (no prunning), a scalar indicating a common hp or
a character indicating the column in \code{data} that corresponds to every tree hp. Default is NULL.}

\item{hm}{Merchandible heigth. Could be NULL, a scalar indicating a common hm for all trees or
a character indicating the column name in \code{data} that corresponds to every tree commercial heigth. Default is NULL}

\item{stump}{Stump height (in m). Default is 0.1.}

\item{lseg0}{Segment size (in m) for the volume estimation using the trapezoid aproximation. Default is 0.2 (NOT IMPLEMENTED).}

\item{hdr_range}{A vector of c(min, max) ht/dbh ratio. This will be used to validate the incoming data. Trees out of this range will be removed.
Default is c(0.5, 2.5) (NOT IMPLEMENTED).}

\item{dbh_range}{A vector of c(min, max) dbh range. This will be used to validate the incoming data. Trees out of this range will be removed.
Default is c(10, 60) (NOT IMPLEMENTED).}

\item{ht_range}{A vector of c(min, max) ht range. This will be used to validate the incoming data. Trees out of this range will be removed.
Default is c(10, 60) (NOT IMPLEMENTED).}

\item{model}{Name of the taper function to use. See details in \code{\link[fpfi3]{taper_selec}} for more information.}

\item{volform}{Volume formula to use. Currently log volumes are calculated using the Smalian formula ("smalian"),
Newton formula ("newton"), geometric formula ("geometric"), Hoppus formula ("hoppus"), and JAS formula ("jas").
See \code{\link[fpfi3]{volform_selec}} for more details. Default is "smalian".}

\item{buck}{A string indicating the type of bucking to be used; options are 'lmak' or 'lpBuck'. Default is 'lmak'.}

\item{grade_wd}{If TRUE, logs are graded after applying the discount to the middle girth. Default is TRUE.}

\item{mg_disc}{Discount in cm to be applied to the middle girth of the logs. Default is 0 (no discount).}

\item{length_disc}{Discount in cm to be applied to the merchantable lengths of the logs. Default is 0 (no discount).}

\item{output}{Option 'logs' will return all the logs obtained for each tree without summarizing by log class name;
"summary" will also return a summary of the total volume for each log assortment at tree level (including logs). Default is "summary".}

\item{byClass}{If TRUE, and \code{output = 'summary'}, the data will be aggregated using the class (column \code{class} in
the \code{products} table), instead of the log name (column \code{name}). Default is FALSE.}
}
\value{
A list with one or two data.table data.frames. It depends on \code{output}.
}
\description{
simulates a stem bucking process based on a list of products using lmak or lp.
}
\details{
To be completed.
}
\examples{
set.seed(87)

demo_products
taper_coeffs
trees <- data.frame(id=1:50, dbh=runif(50)*5 + 35, ht=runif(50)*5 + 30, hp=10)
ans <- DoBucking(trees, taper_coeffs, demo_products, id="id", dbh='dbh', ht='ht',
                 hp='hp', stump=0.15, model="dss", volform="smalian", buck="lmak",
                 grade_wd=TRUE, mg_disc = 0, length_disc = 0, output = "summary",
                 byClass = FALSE)
ans$summary
ans$logs

}
\seealso{
\code{\link[fpfi3]{taper_selec}}, \code{\link[fpfi3]{volform_selec}}, \code{\link[fpfi3]{lmak}}
and \code{\link[fpfi3]{lpBuck}}.
}
\author{
Ãlvaro Paredes, \email{alvaro.paredes@fpfi.cl}, Christian Mora, \email{crmora@fibraconsult.cl}
}
