% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_thin.R
\name{DoThinning}
\alias{DoThinning}
\title{Perform a thinning}
\usage{
DoThinning(species, region, bat, Dgt, Hdt, ntrees, nout, age0 = 0,
  dist = "normal", sd0 = 0, thin_type = "low",
  thin_coefficient = 1, by_class = 1, thin_output = FALSE)
}
\arguments{
\item{species}{A character indicating one of the species name. The function is case insensitive.}

\item{region}{A character indicating one of the region names (depending on the species). The function is case insensitive. Default is NULL.}

\item{bat}{A number indicating the starting average basal area (per ha).}

\item{Dgt}{A number indicating the starting average DBH/Dg.}

\item{Hdt}{A number indicating the starting Hd.}

\item{ntrees}{A number indicating the initial number of trees (per ha).}

\item{nout}{A number indicating the number of trees to be removed (per ha).}

\item{age0}{A number indicating the age in years.}

\item{dist}{A string, indicating the distribution type to be used to simulate the trees. Options are "normal" or "weibull". Default to normal.
As a improvement, this must be separate for thinning (as a vector with different distributions for every thinning) and to produce trees.}

\item{sd0}{A number indicating the standard deviation. If 0, the 10\% of Dgt will be used. Default to 0.}

\item{thin_type}{A string, indicating the thinning type. Options are low, high, dist (distribution sim) or byFunction
(use the \code{species} and \code{region} to determinate the model to be used). Default to low.}

\item{thin_coefficient}{A positive number, indicating the Dgt removed proportion over the remaining Dgt.
A value of 1 indicates a sistematic thinning following the 'dist' distribution. Values over 1 follow a "high" thinning type and below 1, follow a "low" thinning type.
It will be used only if "dist" is selected as the thin_type mode. Usual values range from 0.9 to 1.1 and valid input range from 0.1 to 1.9.
See \code{Details} for more information. Default is 1.}

\item{by_class}{A number, indicating the class interval for the DBH. If 0, the individual density method will be used. Default is 1.}

\item{thin_output}{A boolean. If TRUE, additional output will be computed (frequency of trees by dbh and ht class that are thinned).}
}
\value{
A list with the BA and Hd, and thinning output if thin_output is TRUE.
}
\description{
Calculate the thinning outputs of a single stand.
}
\details{
When \code{thin_coefficient} is out of the range 0.9 to 1.1, it usually produces thinning trees that do not exist.
Inside this function, there's an implementation to correct this, using the \code{\link[fpfi3]{ThinningFit}} function,
so it's safe to use values out of that ranges.
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
