% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{EstimateDBHSd}
\alias{EstimateDBHSd}
\title{Estimate the standard deviation of the DBH}
\usage{
EstimateDBHSd(species, region, SI, N0, ages, sdi, qmds, Alt)
}
\arguments{
\item{species}{A character indicating one of the species name. The function is case insensitive.}

\item{region}{A character indicating one of the region names (depending on the species). The function is case insensitive. Default is NULL.}

\item{SI}{A scalar numeric, indicating the Site Index.}

\item{N0}{A numeric scalar, indicating the initial number of trees per hectare.}

\item{ages}{A numeric vector, indicating the ages t-1 and t, at which the sd would be estimated.
The initial age (at which \code{sdi} is computed), must be the first age (t-1).}

\item{sdi}{A numeric scalar, indicating the initial standard deviation (at t-1).}

\item{qmds}{A numeric vector, indicating the initial quadatric mean diameter at t-1 and t.}

\item{Alt}{The altitude in meters. Used by some functions only. Default is 100.}
}
\value{
A numeric scalar or vector (depends on the input) with the estimated standard deviation.
}
\description{
Computes an estimation of the standard deviation of the DBH, based on
}
\details{
To be completed.
}
\examples{
1+1
}
\seealso{
Other Internals: \code{\link{Dispatcher}},
  \code{\link{GetBAAfterThinning}},
  \code{\link{ProjectBasalArea}},
  \code{\link{ProjectDominantHeight}},
  \code{\link{ProjectNTrees}}, \code{\link{ProjectVolume}},
  \code{\link{SiteIndex}}, \code{\link{availableSpecies}}
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
\concept{Internals}
