% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{fpfi3}
\alias{fpfi3}
\alias{fpfi3-package}
\title{fpfi3: Forestry toolbox for simulation.}
\description{
Growth models, Taper Models, Volume functions and Bucking (with/out LP). Include some plot functions to visualize data.
}
\section{Bucking}{

Functions
}

\section{Plots}{

Functions
}

\section{Simulation}{

Functions
}

\section{Workflow}{

Functions
}

