% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{FullSimulation}
\alias{FullSimulation}
\title{Run a full Growing simulation.}
\usage{
FullSimulation(in_data, age_range, distribution, distribution_thinning,
  species, model, volform, include_thinning = TRUE, byClass = FALSE,
  way = "lmak", region = NULL, stump = 0.15, mg_disc = 0,
  length_disc = 0)
}
\arguments{
\item{in_data}{A list with three data.frames, in the form list(df=initial conditions, params=parameteres, prods=products)}

\item{age_range}{A vector with the c(min, max) age range to be used in the cutting simulation.}

\item{distribution}{See \code{dist} in \code{\link[fpfi3]{SimulateTrees}}.}

\item{distribution_thinning}{See \code{dist} in \code{\link[fpfi3]{DoThinning}}.}

\item{species}{A String indicating the species. Options are 'eucalyptus', 'pine', 'teak'.}

\item{model}{See \code{\link[fpfi3]{taper_selec}}}

\item{volform}{See \code{\link[fpfi3]{DoBucking}}}

\item{include_thinning}{if TRUE, the yield by production at every thinning will be added to the output. Default is TRUE.}

\item{byClass}{If TRUE, and \code{output = 'summary'}, the data will be aggregated using the class (column \code{class} in
the \code{products} table), instead of the log name (column \code{name}). Default is FALSE.}

\item{way}{A string indicating the taper way to compute the volume production. Options are "lmak" or "lpBuck". Defualt is \code{lmak}.
See \code{\link{lpBuck}} and \code{\link{lmak}} for more information.}

\item{region}{An optional string value, indicating the region.
If NULL, global parameters for the species will be used. Options are listed in the \code{details} section. Default is NULL.}

\item{stump}{Stump height (in m). Default is 0.1.}

\item{mg_disc}{Discount in cm to be applied to the middle girth of the logs. Default is 0 (no discount).}

\item{length_disc}{Discount in cm to be applied to the merchantable lengths of the logs. Default is 0 (no discount).}
}
\value{
An S3 object of class \code{FullSimulation}. This object is basically a list, consisting in:
\itemize{
\item \code{simulation}: A \code{data.table} data.frame with the growth simulation and others computed indices (as columns).
\item \code{taper}: A list with two levels of details.
\itemize{
\item \code{log_level}: A list with the individuals logs by tree, separated by \code{harvest} and \code{thinning}. Each table have:
\itemize{
\item \code{harvest}:
\item \code{thinning}:
}
\item \code{tree_level}: A list with the individual tree level detail separated by \code{harvest} and \code{thinning}. Each table have:
\itemize{
\item \code{harvest}:
\item \code{thinning}:
}
\item \code{stand_level}: A list with the summary of \code{tree_level}. It usually used for the LP input and it's separated \code{harvest} and \code{thinning}:
\itemize{
\item \code{harvest}:
\item \code{thinning}:
}
}
\item \code{input}: A list with the tree inputs:
\itemize{
\item \code{initial_conditions}: A \code{data.table} data.frame with the initial condictions for every stand.
\item \code{taper_coefficients}: A named vector, with the coefficients for the taper function \code{\link[fpfi3]{taper_selec}}.
\item \code{products}: A \code{data.table} data.frame with the initial products.
}
\item \code{growth}: A list, returned by \code{\link[fpfi3]{BatchGrowing}}.
\item \code{cl}: A call returned by \code{match.call}, with the initial parameters of the \code{FullSimulation} call.
}
}
\description{
Create a full simulation from an initial condition input and returning the full volume simulation (harvest and thinning).
The function will start simulating the growth, then the taper and bucking for the harvest (and optionally, the thinning) volume, in order
to generate the aggregated yield tables.
}
\details{
This function is a wrapper for several individual functions. In first place, an instance of \code{\link[fpfi3]{BatchGrowing}}
is used to generate all the simulated properties for every stand, at all the desired ages. That simulation will be used to peform the taper and
the bucking/grading by the \code{\link[fpfi3]{DoBucking}}, for harvest and thinning (depending on the condition set in \code{include_thinning}.
Ultimately, union/join/merge operations will be used to mantain the consistency between the input tables and the output tables. Additionally
information will be added to the output.
}
\examples{
library(fpfi3)

# Let's use the data loaded inside this package:
initial_conditions  # See ?initial_conditions for a detailed description.
taper_coeffs  # See ?taper_coeffs for a detailed description.
demo_products  # See ?demo_products for a detailed description.

# Running the simulation
input_data <- list(df=initial_conditions, params=taper_coeffs, prods=demo_products)
sims <- FullSimulation(in_data = input_data,
                       age_range = c(15,30), distribution = "normal",
                       distribution_thinning = "normal", include_thinning = TRUE,
                       species="eucalyptus_grandis", model = "dss", volform = "smalian",
                       way = "lmak", region = 'uruguay_solid',
                       stump=0.15, mg_disc=0, length_disc=0)

}
\seealso{
\code{\link[fpfi3]{BatchGrowing}} and \code{\link[fpfi3]{DoBucking}}, for a detailed explanation of the internal methods. To see which
species are available \code{\link[fpfi3]{availableSpecies}}.

To visualize results, \code{\link[fpfi3]{plotHist}}.
}
\author{
Ãlvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
