% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{GetBAAfterThinning}
\alias{GetBAAfterThinning}
\title{Project the Basal Area (BA) after the thinning action.}
\usage{
GetBAAfterThinning(species, region, BA, N1, N2, Hd)
}
\arguments{
\item{species}{A character indicating one of the species name. The function is case insensitive.}

\item{region}{A character indicating one of the region names (depending on the species). The function is case insensitive. Default is NULL.}

\item{BA}{A scalar numeric indicating the initial BA (previous to thinning).}

\item{N1}{Number of trees per hectarea previous to the thinning.}

\item{N2}{Number of trees per hectarea after the thinning.}

\item{Hd}{A scalar numeric indicating the dominant height.}
}
\value{
A scalar numeric.
}
\description{
This function project the basal area after a thinning, using different parameters and structure
depending on the species and region.
}
\details{
Models were extracted from:
\itemize{
\item Eucalyptus:
\itemize{
\item Uruguay: to be completed.
}
\item Pinus Radiata:
\itemize{
\item New Zealand: to be completed.
}
\item Pinus Taeda:
\itemize{
\item Uruguay: to be completed.
}
}
}
\examples{
1+1
}
\seealso{
Other Internals: \code{\link{Dispatcher}},
  \code{\link{EstimateDBHSd}},
  \code{\link{ProjectBasalArea}},
  \code{\link{ProjectDominantHeight}},
  \code{\link{ProjectNTrees}}, \code{\link{ProjectVolume}},
  \code{\link{SiteIndex}}, \code{\link{availableSpecies}}
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
\concept{Internals}
