% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{Growing}
\alias{Growing}
\title{Main Growing function}
\usage{
Growing(Age0, Hd0, N0, Dg0, BA0, Vt0, species, Dg0_sd = 0, years = 30,
  region = NULL, thinningAge = NULL, n.afterThin = NULL,
  thin_type = NULL, thin_coefficient = NULL, ...)
}
\arguments{
\item{Age0}{A number indicating the starting average age.}

\item{Hd0}{A number indicaring the starting tree dominant height.}

\item{N0}{A number indicating the starting number of trees per ha.}

\item{Dg0}{A number indicating the starting average DBH/Dg.}

\item{BA0}{A number indicarting the starting average basal area (per ha).}

\item{Vt0}{A numner indicating the starting average total volume (per ha).}

\item{species}{A string indicating the species.}

\item{Dg0_sd}{A number indicating the actual standard deviation of \code{Dg0}. Default is 0 that's mean that an attempt to predict
the standard deviation (depending on the species) will be conducted, using \code{\link[fpfi3]{EstimateDBHSd}}.}

\item{years}{A number indicating the number of years to simulate. Default to 30.}

\item{region}{A string indicating the region. Default is \code{NULL}.}

\item{thinningAge}{A vector containing the thinning ages as numbers (in increasing order). Default to no thinning.}

\item{n.afterThin}{A vector containing the number of trees after the thinning.
(related to thinningAge; must have the same length). Default to no thinning.}

\item{thin_type}{See \code{\link[fpfi3]{DoThinning}} for more details.}

\item{thin_coefficient}{See \code{\link[fpfi3]{DoThinning}} for more details.}

\item{...}{Additional arguments passed to \code{\link[fpfi3]{DoThinning}} function.}
}
\value{
A data.frame with the Growing
}
\description{
Growings function: from initial conditions, compute Dg, BA, N, Vt and thinning action until certain time period (year)
}
\details{
To be completed.
}
\examples{
1+1
}
\seealso{
\code{\link[fpfi3]{EstimateDBHSd}} and \code{\link[fpfi3]{DoThinning}}.
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
