% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{initial_conditions}
\alias{initial_conditions}
\title{Initial Conditions}
\format{A \code{data.frame}, \code{data.table} with 20 variables:
\itemize{
\item{\code{id: }}{A unique numeric identificator.}
\item{\code{macrostand: }}{A character indicating the stand name.}
}}
\usage{
initial_conditions
}
\description{
An example of the initial conditions for 5 stands.
}
\keyword{internal}
