% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmak.R
\name{lmak}
\alias{lmak}
\title{Stem bucking simulation and merchantable volume calculation}
\usage{
lmak(treesDiam, treesDat, products, hp = NULL, lseg0 = 0.2,
  volform = "smalian", grade_wd = FALSE, mg_disc = NULL,
  length_disc = NULL)
}
\arguments{
\item{treesDiam}{A data.table or matrix, where every row corresponds to a tree and every column to a diameter at a given bin.}

\item{treesDat}{A data.table dataframe, with colnames idseq, ht and dbh at least. Every row corresponds to a tree (in the same order as treeVol).}

\item{products}{A data.frame of 6 columns and n rows, with product class information. The order is important,
it will mark the priority (first products go before the others in Lmak):
\itemize{
\item Column 1 is the small-end diameter ("diameter").
\item Column 2 is product length ("length") which has to be the same for all poducts.
\item Column 3 is the product value ("value").
\item Column 4 is the product name ("name").
\item Column 5 is the log type ("log_type"), that indicate if the log is prunned ("p") or non-prunned ("u").
\item Column 6 is the log class ("class"), that is used to aggregate or classify the data.
}}

\item{hp}{Prunning height. Could be NULL (no prunning), a scalar indicating a common hp or
a character indicating the column in data that corresponds to every tree hp. Default is NULL.}

\item{lseg0}{Segment size (in m) for the volume estimation using the trapezoid aproximation. Default is 0.2 (NOT IMPLEMENTED).}

\item{volform}{Volume formula to use. Currently log volumes are calculated using the Smalian formula ("smalian"),
Newton formula ("newton"), geometric formula ("geometric"), Hoppus formula ("hoppus"), and JAS formula ("jas").
See \code{\link{volform_selec}} for more details. Default is "smalian".}

\item{grade_wd}{If TRUE, logs are graded after applying the discount to the middle girth. Default is TRUE.}

\item{mg_disc}{Discount in cm to be applied to the middle girth of the logs. Default is NULL (no discount).}

\item{length_disc}{Discount in cm to be applied to the merchantable lengths of the logs. Default is NULL (no discount).}
}
\value{
A data.table dataframe.
}
\description{
Lmak simulates a stem bucking process based on a list of products of fixed(same) length.
}
\details{
This functions is designed to be used by \code{\link[fpfi3]{DoBucking}}. Be carefull to give all the arguments in the proper as no major
checks will be performed (all checks and data preparation are made by the \code{\link[fpfi3]{DoBucking}} function).
}
\examples{
1+1

}
\author{
Christian Mora, \email{crmora@fibraconsult.cl}, Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
