% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp.R
\name{lpBuck}
\alias{lpBuck}
\title{Stem bcking simulation and merchantable volume calculation by linear model}
\usage{
lpBuck(treesDiam, treesDat, products, hp = NULL, lseg0 = 0.2,
  volform = "smalian", grade_wd = FALSE, mg_disc = NULL,
  length_disc = NULL)
}
\arguments{
\item{treesDiam}{A data.table or matrix, where every row corresponds to a tree and every column to a diameter at a given bin.}

\item{treesDat}{A data.table dataframe, with colnames idseq, ht and dbh at least. Every row corresponds to a tree (in the same order as treeVol).}

\item{products}{A data.frame of 4 columns and n rows, with product class information.
Column 1 is the small-end diameter ("diameter"), column 2 is product length ("length") that could be different for each one,
columnn 3 is the value ("value") of the log and column 4 is the product name ("name").
The order is not relevant, as logs with more value will be selected automatically by the solver.}

\item{hp}{Prunning height. Could be NULL (no prunning), a scalar indicating a common hp or
a character indicating the column in data that corresponds to every tree hp. Default is NULL (NOT IMPLEMENTED).}

\item{lseg0}{Segment size (in m) for the volume estimation using the trapezoid aproximation. Default is 0.2 (NOT IMPLEMENTED).}

\item{volform}{Volume formula to use. Currently log volumes are calculated using the Smalian formula ("smalian"),
Newton formula ("newton"), geometric formula ("geometric"), Hoppus formula ("hoppus"), and JAS formula ("jas").
See \code{\link{volform_selec}} for more details. Default is "smalian".}

\item{grade_wd}{If TRUE, logs are graded after applying the discount to the middle girth. Default is TRUE.}

\item{mg_disc}{Discount in cm to be applied to the middle girth of the logs. Default is NULL (no discount).}

\item{length_disc}{Discount in cm to be applied to the merchantable lengths of the logs. Default is NULL (no discount).}
}
\value{
A data.table dataframe.
}
\description{
lpBuck simulates a stem bucking process based on a list of products of differents lengths, using an integer lineal model in the process.
}
\details{
This functions is designed to be used by \code{\link[fpfi3]{DoBucking}}. Be carefull to give all the arguments in the proper as no major
checks will be performed (all checks and data preparation are made by the \code{\link[fpfi3]{DoBucking}} function).
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
