% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotHist}
\alias{plotHist}
\title{Plot histogram}
\usage{
plotHist(obj, ids = NULL, macrostands = NULL, plot = TRUE,
  subset = 25, points.after = 0)
}
\arguments{
\item{obj}{A \code{FullSimulation} object generated by the \code{\link[fpfi3]{FullSimulation}} function.}

\item{ids}{A vector of idg to plot. Is the 'idg' column of the "main" parameter in the list returned by the BatchGrowing function. Default is NULL (first 9 ploted)}

\item{macrostands}{A vector of names.}

\item{plot}{if TRUE, the plot will be displayed. Default is TRUE.}

\item{subset}{A number, indicating the maximum number of plots to display (to avoid saturation). Default value is 25 (5x5 grid).}

\item{points.after}{An integer, indicating how many years after the thinning to plot (usually to compare). Default is 0 (no plot).}
}
\value{
A plot, where grey is the pre-thinned distribution, blue is the thinned distribution and red is the after-thinning distribution of trees.
If \code{points.after != 0}, the distribution will appear in green.
}
\description{
Plot the histogram of the thinned vs harvested vs total.
}
\details{
A maximum of 16 plots are allowed in one single call.
}
\examples{
library(fpfi3)

# Let's use the data loaded inside this package:
initial_conditions  # See ?initial_conditions for a detailed description.
taper_coeffs  # See ?taper_coeffs for a detailed description.
demo_products  # See ?demo_products for a detailed description.

# Running the simulation
input_data <- list(df=initial_conditions, params=taper_coeffs, prods=demo_products)
sims <- FullSimulation(in_data = input_data,
                       age_range = c(15,30), distribution = "normal",
                       distribution_thinning = "normal", include_thinning = TRUE,
                       species="eucalyptus_grandis", model = "dss", volform = "smalian",
                       way = "lmak", region = 'uruguay_solid')

plotHist(sims)  # plot all
plotHist(sims, macrostands = c("A1", "A2"))
plotHist(sims, macrostands = c("A1", "A2"), points.after=1)

}
\seealso{
\code{\link[fpfi3]{FullSimulation}}
}
\author{
Ãlvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
