% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print.FullSimulation}
\alias{print.FullSimulation}
\title{FullSimulation Printing Options}
\usage{
\method{print}{FullSimulation}(x, ...)
}
\arguments{
\item{x}{A \code{FullSimulation} object (generated by the \code{\link[fpfi3]{FullSimulation}} function.)}

\item{...}{Other arguments ultimately passed to format.}
}
\description{
\code{print.FullSimulation} allow a better visualization that the normal \code{print}.
It shows a summary of the available information and not all the list components in a single print.
}
\details{
To be completed.
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
\keyword{internal}
