% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{ProjectBasalArea}
\alias{ProjectBasalArea}
\title{Project the Basal Area (BA)}
\usage{
ProjectBasalArea(species, region, BA1, Age1, Age2, SI, N0, Alt = 100)
}
\arguments{
\item{species}{A character indicating one of the species name. The function is case insensitive.}

\item{region}{A character indicating one of the region names (depending on the species). The function is case insensitive. Default is NULL.}

\item{BA1}{A scalar numeric indicating the initial BA.}

\item{Age1}{A scalar numeric indicating the Age when \code{BA1} is measured.}

\item{Age2}{A scalar integer at which the projection is desired.}

\item{SI}{A scalar numeric indicating the Site Index.}

\item{N0}{A scalar numeric indicating the number of trees when \code{BA1} is measured.}

\item{Alt}{The altitude in meters. Used by some functions only. Default is 100.}
}
\value{
A scalar numeric.
}
\description{
This function project the basal area at any desired age, using different parameters and structure
depending on the species and region.
}
\details{
Models were extracted from:
\itemize{
\item Eucalyptus:
\itemize{
\item Uruguay: to be completed.
}
\item Pinus Radiata:
\itemize{
\item New Zealand: to be completed.
}
\item Pinus Taeda:
\itemize{
\item Uruguay: to be completed.
}
}
}
\examples{
1+1
}
\seealso{
Other Internals: \code{\link{Dispatcher}},
  \code{\link{EstimateDBHSd}},
  \code{\link{GetBAAfterThinning}},
  \code{\link{ProjectDominantHeight}},
  \code{\link{ProjectNTrees}}, \code{\link{ProjectVolume}},
  \code{\link{SiteIndex}}, \code{\link{availableSpecies}}
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
\concept{Internals}
