% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{ProjectVolume}
\alias{ProjectVolume}
\title{Project the volume.}
\usage{
ProjectVolume(species, region, BA, Hd)
}
\arguments{
\item{species}{A character indicating one of the species name. The function is case insensitive.}

\item{region}{A character indicating one of the region names (depending on the species). The function is case insensitive. Default is NULL.}

\item{BA}{A scalar numeric indicating the basal area.}

\item{Hd}{A scalar numeric indicating the dominant height}
}
\value{
A scalar numeric.
}
\description{
This function project the volume at any desired age, using different parameters and structure
depending on the species and region.
}
\details{
Models were extracted from:
\itemize{
\item Eucalyptus:
\itemize{
\item Uruguay: document
}
\item Pinus Radiata:
\itemize{
\item New Zealand: document
}
\item Pinus Taeda:
\itemize{
\item Uruguay: document
}
}
}
\examples{
1+1
}
\seealso{
Other Internals: \code{\link{Dispatcher}},
  \code{\link{EstimateDBHSd}},
  \code{\link{GetBAAfterThinning}},
  \code{\link{ProjectBasalArea}},
  \code{\link{ProjectDominantHeight}},
  \code{\link{ProjectNTrees}}, \code{\link{SiteIndex}},
  \code{\link{availableSpecies}}
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
\concept{Internals}
