% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{SimulateTrees}
\alias{SimulateTrees}
\title{Simulate trees}
\usage{
SimulateTrees(ntrees, by_class, dist, Dgt = 0, Hdt = 0, sd0 = 0,
  BA = 0, age0 = 0, full = FALSE, ...)
}
\arguments{
\item{ntrees}{A number indicating the initial number of trees (per ha)}

\item{by_class}{A number, indicating the class interval for the DBH. If 0, the individual density method will be used. Default is 1.}

\item{dist}{A string, indicating the distribution type to be used to simulate the trees. Options are "normal" or "weibull". Default to normal.
As a improvement, this must be separate for thinning (as a vector with different distributions for every thinning) and to produce trees.}

\item{Dgt}{A number indicating the starting average DBH/Dg}

\item{Hdt}{A number indicating the starting Hd}

\item{sd0}{A number indicating the standard deviation. If 0, an attempt to simulate the sd will be conducted using \code{\link[fpfi3]{EstimateDBHSd}}
, otherwise the 10\% of Dgt will be used. Default to 0.}

\item{BA}{A number indicating the Basal Area}

\item{age0}{A number indicating the stand age}

\item{full}{A boolean. If TRUE, the height and other variables will be computed with dbh. Default is False.}

\item{...}{Additional arguments passed to \code{\link[fpfi3]{SimulateTreesHeigth}}}
}
\value{
An object (list) of the class \code{TreeSim} with several attributes
}
\description{
Auxiliary function for DoThinning, to simulate the trees given a certain distribution and other parameters.
}
\details{
To be completed.
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
