% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{SimulateTreesHeigth}
\alias{SimulateTreesHeigth}
\title{Simulate tree height.}
\usage{
SimulateTreesHeigth(species, region, trees, Hd, age, BA, SI, Alt = 100)
}
\arguments{
\item{species}{A character indicating one of the species name. The function is case insensitive.}

\item{region}{A character indicating one of the region names (depending on the species). The function is case insensitive. Default is NULL.}

\item{trees}{A numeric vector indicating the dbh of individual trees.}

\item{Hd}{Dominant height}

\item{age}{Age (in years)}

\item{BA}{Basal Area.}

\item{SI}{Site Index.}

\item{Alt}{Altitude over the sea level in meters. Default is 100.}
}
\value{
A list with Ht(trees), d100, tph, alpha and beta
}
\description{
Function to simulate the individual trees height, from a list of individual dbhs
}
\details{
Note that use of certain parameters depends on the species and region.
}
\examples{
1+1
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
