% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{SiteIndex}
\alias{SiteIndex}
\title{Calculate the site index of one place}
\usage{
SiteIndex(species, region, Hd1, Age1, Alt = 100)
}
\arguments{
\item{species}{A character indicating one of the species name. The function is case insensitive.}

\item{region}{A character indicating one of the region names (depending on the species). The function is case insensitive. Default is NULL.}

\item{Hd1}{Initial Dominant Height.}

\item{Age1}{A scalar numeric indicating the Age at wich \code{Hd1} is measured.}

\item{Alt}{The altitude in meters. Used by some functions only. Default is 100.}
}
\value{
A scalar numeric.
}
\description{
This function project the dominant height at any desired age, using different parameters and structure
depending on the species and region.
}
\details{
This function is a parser for the previous function, using defined ages by species and region.
For all species is age 10, except for pinus Radiata which is 20.

See \code{\link[fpfi3]{ProjectDominantHeight}} for more information.
}
\examples{
1+1
}
\seealso{
Other Internals: \code{\link{Dispatcher}},
  \code{\link{EstimateDBHSd}},
  \code{\link{GetBAAfterThinning}},
  \code{\link{ProjectBasalArea}},
  \code{\link{ProjectDominantHeight}},
  \code{\link{ProjectNTrees}}, \code{\link{ProjectVolume}},
  \code{\link{availableSpecies}}
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
\concept{Internals}
