% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taper_coeffs}
\alias{taper_coeffs}
\title{Coefficients}
\format{See \code{\link[fpfi3]{taper_selec}}.}
\usage{
taper_coeffs
}
\description{
An example of the taper coefficients.
}
\keyword{internal}
