% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taper_func.R
\name{taper_selec}
\alias{taper_selec}
\title{Taper form selector.}
\usage{
taper_selec(func_name, x, coeffs, cut_ht, dbh, ht, hm = NULL)
}
\arguments{
\item{func_name}{A string indicating the model to be used. See \code{Details} for more information.}

\item{x}{A data.table data frame with the at least, with the dbh and ht. Every row is a different tree}

\item{coeffs}{A vector with the coefficients for the model to be used.}

\item{cut_ht}{A vector with the desired cut heights.}

\item{dbh}{A string, indicating the column name in \code{x} that containts the Diameter at Breast Height (dbh) of every tree.}

\item{ht}{A string, indicating the column name in \code{x} that containts the heigth (ht) of every tree.}

\item{hm}{A string or a scalar number. In the first case (string), it means the column in \code{x} that shows every tree commercial height.
If scalar number, it will be used as a common commercial heigth for all the trees. Default is NULL, which means that commercial heigth will not be used.}
}
\value{
A data.table dataframe with the tree diameter per every \code{cut_ht} and tree (rows)
}
\description{
Select the desired function.
}
\details{
The full list of implemented models are below. Default is "dss":
\enumerate{
\item \code{bruce}: the Bruce's model for pine.
See the \href{http://www.ingentaconnect.com/content/saf/fs/1968/00000014/00000003/art00021}{documentation}.
\item \code{cao}: the modified segmented Max-Burkhart model developed by Cao.
\item \code{cao2}: the modified segmented Max-Burkhart model implemented by FPFI.
See the \href{http://www.ingentaconnect.com/content/saf/fs/2015/00000061/00000002/art00002}{documentation}.
\item \code{dss}: the modified segmented Max-Burkhart model implemented in DSS grandis (SAG grandis model).
See the \href{http://www.ainfo.inia.uy/digital/bitstream/item/2876/1/15630021107224844.pdf}{documentation}.
\item \code{fang}: the Fang-Border-Bailey model.
See the \href{http://www.ingentaconnect.com/contentone/saf/fs/2000/00000046/00000001/art00002}{documentation}.
\item \code{fds}: the 7th degree polynomial model developed by Granflor for FDS.
\item \code{garay}: the Garay's taper model.
\item \code{mb}: the segmented Max-Burkhart polynomial model.
See the \href{http://www.ingentaconnect.com/content/saf/fs/1976/00000022/00000003/art00011}{documentation}.
\item \code{mb_chile}: the segmented Max-Burkhart polynomial model (adapted for Chile).
See model 17 in the \href{http://www.redalyc.org/pdf/629/62945379008.pdf}{documentation}.
\item \code{mm}: the Muhairwe's model modified by Methol.
See the \href{http://www.scielo.edu.uy/pdf/agro/v18n2/v18n2a06.pdf}{documentation}.
\item \code{parresol}: the Thomas and Parresol model for pine.
See the \href{http://www.nrcresearchpress.com/doi/abs/10.1139/x91-157?journalCode=cjfr#.WV7vylGQwqM}{documentation}.
\item \code{poly5}: the 5th degree polynomial model.
See the \href{http://cerne.ufla.br/ojs/index.php/CERNE/article/view/604/513}{documentation}.
\item \code{prad}: the 6th degree polynomial model developed by Granflor for PRAD.
\item \code{sispinus}: the 4th degree polynomial model used in SisPinus.
\item \code{warner}: See the \href{http://dx.doi.org/10.1016/j.anres.2016.04.005}{documentation}.
}
}
\author{
Álvaro Paredes, \email{alvaro.paredes@fpfi.cl}, Christian Mora, \email{crmora@fibraconsult.cl}
}
