% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_thin.R
\name{ThinningFit}
\alias{ThinningFit}
\title{Adjust the thinning output to fit the initial available trees.}
\usage{
ThinningFit(treeSim, thinSim)
}
\arguments{
\item{treeSim}{an object of the class \code{TreeSim} returned by the \code{\link[fpfi3]{SimulateTrees}} function, representing the initial trees (previous to thinning).}

\item{thinSim}{an object of the class \code{TreeSim} returned by the \code{\link[fpfi3]{SimulateTrees}} function, representing the thinned trees.}
}
\value{
A list with the adjusted thinning
}
\description{
The thinned diameter class will be forced to fit the maximum available trees per diameter class. Some discrepancies could happen
when using a distributed thinning (option \code{'dist'}) in the \code{\link[fpfi3]{DoThinning}} function and the thinned trees could result in
removed trees that do not exist. This functions aims to relocate the non existing trees into the initial distribution.
If thinned average dbh is lower than initial average dbh, it means it's a 'lower' thinning and the non-existing trees will be
sequencially moved to the next category, starting from the lower diameters; otherwhise, when
the thinned average dbh is greater than the initial average dbh, the starting point will be the higher diameters.
}
\details{
Note that this function was implemented to be used inside \code{\link[fpfi3]{DoThinning}}.
}
\examples{
# Usinng simulated trees
library(fpfi3)

thin_coef <- 0.8
Dg <- 21.3
sd <- 2.75
trees <- SimulateTrees(ntrees=432, by_class=1, dist="normal", Dgt=Dg, Hdt=21.8, sd0=sd,
BA=15.4, age0=7, full=TRUE, species='eucalyptus_grandis', region='uruguay_solid')
thinned_trees <- SimulateTrees(ntrees=182, by_class=1, dist='normal', Dgt=Dg * thin_coef,
Hdt=21.8, sd0=.1 * Dg * thin_coef, BA=15.4, age0=7, full=TRUE,
 species='eucalyptus_grandis', region='uruguay_solid')

ndata <- ThinningFit(trees, thinned_trees)

# Visual check of the results
plot(trees$class_tree[, c("trees", "Freq")],
type="b", pch=16, xlim=range(ndata$class_tree$trees), xaxt="n")
axis(1, ndata$class_tree$trees)
abline(v=ndata$class_tree$trees, col=rgb(.7, .7, .7, .5))
lines(thinned_trees$class_tree[, c("trees", "Freq")], col="darkgreen")
lines(ndata$class_tree[, c("trees", "Freq")], col="red")
legend('topright', legend=c("Initial trees", "Initial thinning", "Fitted thinning"),
col=c('black', 'darkgreen', 'red'), lty=1)

}
\author{
Ãlvaro Paredes, \email{alvaro.paredes@fpfi.cl}
}
