% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vol_func.R
\name{volform_selec}
\alias{volform_selec}
\title{Volume form selector.}
\usage{
volform_selec(func_name)
}
\arguments{
\item{func_name}{A string indicating the model to be used. See Details for more information.}
}
\value{
The selected function model.
}
\description{
Select the desired function.
}
\details{
There are 5 models available. They use different parts of the log to make the volume calculation, based on the large end diameter (LED),
middle diameter (MID) or small end diameter (SED). See the pdf version of this help if the equations are not displayed correctly.

\code{geometric}: Usually used in teak. The volume calculations uses the log MID.

\deqn{V(m^3) = MID^2 \times L \times 6.1685^{-5}}
where:
\itemize{
\item MID = Middle Diameter in cm.
\item L = length in m.
}

\code{hoppus}: Usually used in teak. The volume calculations uses the log MID.

\deqn{V(m^3) = \frac{MID^2}{40000} \times pi \times L}
where:
\itemize{
\item MID = Middle Diameter in cm.
\item L = length in m.
}

\code{smalian}: The volume calculations uses the log LED and SED.

\deqn{V(m^3) = \frac{LED^2 + SED^2}{80000} \times pi \times L}
where:
\itemize{
\item LED = Large End Diameter in cm.
\item SED = Small End Diameter in cm.
\item L = length in m.
}

\code{newton}: The volume calculations uses the log LED, MID and SED.

\deqn{V(m^3) = \frac{LED^2 + 4 \times MID^2 + SED^2}{240000} \times pi \times L}
where:
\itemize{
\item LED = Large End Diamater in cm.
\item MID = Middle Diameter in cm.
\item SED = Small End Diameter in cm.
\item L = length in m.
}

\code{jas}: Japanese Agricultural Standard (JAS). The volume calculations uses the log SED.

If the log is less than 6 meters:
\deqn{V(m^3) = \frac{SED^2*L}{10000}}

For logs equal or greater than 6 meters:
\deqn{V(m^3) = \frac{SED + (\frac{L' - 4}{2})^2 * L}{10000}}
where:
\itemize{
\item SED = Small End Diameter in cm (rounded to the lower even centimeter). No decimal places.
\item L = length in m (rounded to the lower decimeter). One decimal place (some places rounded to the lower even decimeter).
\item L' = length in m (rounded to the lower meter). No decimal places.
}
}
